refactor: overlay 배럴 파일 삭제
