--- src/components/overlay/index.tsx
@@ -1,3 +1,0 @@
-export { default as Toast } from './Toast';
-export { default as FullScreenModal } from './FullScreenModal';
-export { default as Portal } from './Portal';
